Fix for crash during button pushes
One more final reason for intermittent resets, and this one is software
fixable. You are pushing the buttons way too fast.

The issue, is that every time you push the button, or flip a switch, you
cause a processor interrupt. Now, in order for the processor to return
to where it was interrupted from, it has to store that interrupt return
location somewhere. Unfortunately, that somewhere, is in the limited
ram. If the interrupt handler does not finish processing before it
is interrupted, then more ram is consumed by this interrupt. If timed
right, you could very well be filling all of the ram, corrupting what
is in the ram, like current date, time, and all of that stuff. Worse,
this corruption, might cause some loop to lock up somewhere, causing
the watchdog to kick in and reset the processor.

What really needs to happen, if interrupts are to be enabled, the button
interrupts themselves are to be disabled till the processing of that
interrupt is complete.

The alarm switch is another interrupt that really needs to be disabled
while it is processing, for the same reason. Someone could flick it back
and forth fast enough to cause corruption of memory.

I believe I have fixed the issue completely. The solution is to disable
the specific interrupt upon entry, and re-enable it just before exit of
the interrupt handling routine. It is okay to allow all other interrupts
to happen in the mean time, but sometimes, indeed, the switches bounce,
and the interrupt fires very fast as a result.

http://forums.adafruit.com/viewtopic.php?p=62976&sid=b8163ae02059c4337990d4a68350891f#p62976