Rework button handling
Remove ad-hoc bespoke button handling and centralize it to a single
place.  This makes things much simpler for code which wants to act on
buttons, and gives quite a large overall code-size saving.

This models the buttons with a 4-state finite state machine.  The states are:
  OPEN		- button is not being pressed
  CLOSED	- button is being pressed, but hasn't been debounced
  LATCHED	- pressed long enough to be debounced and available for reading
  SAMPLED	- button has been read, but not released or repeated

There are two functions which update the state:
  button_change_intr - called by the interrupt handler
  		       when a button is pressed/released
  button_state_update - called every (approx) ms to update
  		      	time-based state

Debounce is handled by waiting for 10ms before moving a button from
CLOSED to LATCHED; if the button is released in the meantime, it goes
from CLOSED to OPEN.

Repeat is implemented by making the button go from SAMPLED->LATCHED
after an initial timeout and repeat timeouts.  This is only
implemented on button 3 (+), since repeating MENU and SET aren't that
useful.

The full state-machine is in button-fsm.dot

button_sample() returns true if a button is in LATCHED state, and moves
the button to SAMPLED.

button_poll() returns true if the button is LATCHED, but doesn't change the
state.
* * *
Prime button states on initialization.
* * *
Make sure button state isn't cached.
* * *
Two if block with the same test?