Generalized table-driven menu handling.
Rather than having ad-hoc code to implement menus repeated for each
item, replace it with a table-driven approach which make all menu
behaviour centralized.

The main menu is a list of prompts with associated field arrays.

Each field array fits on one display, and shows a mixture of prompts
and values which can be updated.

When entering the main menu, it shows each prompt in turn, selected
with MENU.  When the user pushes SET, it displays the set of fields,
with the first input field highlighted.

Pressing (+) (NEXT in the code) calls the update function on the field
to move to the next value, and refreshes the display.

Pressing SET moves to the next input field, or leaves the menu if it
was the last.
* * *
Brightness is just a number, so just use show_num.
* * *
emit_number_dots no longer used.