Add time-of-date auto-brightness setting.
Replaces brightness menu with two new entries which set "day"
brightness and "night" brightness.

Each has the hour at which day and night start, which are enforced to
be in the morning (midnight-midday) and afternoon (midday-midnight)
respectively.

Old behaviour can be emulated by setting the brightness the same for
both and ignoring the time.