Drift correction infrastructure.
Rather than using overflow on timer/counter, use CTC mode
(clear and interrupt on compare) with double the divider.  We can
then tune how long a second is to within ~1%.  This is too coarse
to apply to every second, but applying it to the first second of an
hour gets an a drift correction of up to ~300ppm in ~2ppm increments.

Ref http://forums.adafruit.com/viewtopic.php?f=41&t=12720