Fix time progression
Using 128 as the baseline for OCR2A makes seconds 1/128th too long,
which loses time at about .5sec/min.

Also, only update OCR2A when necessary (on the hour for drift correction,
and the second later to restore baseline), and wait for ASSR.OCR2AUB to
clear before continuing, as the manual recommends.

Also also, zero TCNT2 when setting time, so that the second starts exactly
when the user sets the time.